fix(long): assign es_type long to all ints (#125)

+++ tube/etl/outputs/es/writer.py
@@ -34,7 +34,7 @@ class Writer(SparkBase):
         es_type = {
             str: 'keyword',
             float: 'float',
-            int: 'integer'
+            int: 'long'
         }
 
         properties = {
